+++ packages/games/src/frontend/fishbowl/globalScreen/StartGame.tsx
@@ -1,12 +1,12 @@
 import { PlayerId } from "@resync-games/api";
-import { Button, DisplayText } from "../../../../../lib/radix";
-import { FishbowlGameConfiguration } from "../../../../backend";
-import { newRound } from "../../stateFunctions/newRound";
+import { Button, DisplayText } from "../../../../lib/radix";
+import { FishbowlGameConfiguration } from "../../../backend";
+import { newRound } from "../stateFunctions/newRound";
 import {
   updateFishbowlGameState,
   useFishbowlDispatch,
   useFishbowlSelector
-} from "../../store/fishbowlRedux";
+} from "../store/fishbowlRedux";
 
 export const StartGame = () => {
   const dispatch = useFishbowlDispatch();
