+++ packages/games/src/frontend/fishbowl/globalScreen/components/ActivePlayerTracker.module.scss
