+++ packages/games/src/frontend/fishbowl/globalScreen/components/ActivePlayerTracker.tsx
@@ -1,6 +1,6 @@
 import { PlayerIcon } from "@/components/player/PlayerIcon";
 import { DisplayText, Flex } from "@/lib/radix";
-import { useFishbowlSelector } from "../../../store/fishbowlRedux";
+import { useFishbowlSelector } from "../../store/fishbowlRedux";
 import { TimerState } from "./TimerState";
 import styles from "./ActivePlayerTracker.module.scss";
 import { getTeamColor } from "@/lib/stableIdentifiers/teamIdentifier";
