+++ packages/games/src/frontend/fishbowl/globalScreen/components/TimerState.module.scss
