+++ packages/games/src/frontend/fishbowl/playerComponents/activePlayer/SomeoneGotIt.tsx
@@ -1,9 +1,9 @@
 import { SlideConfirm } from "@/lib/radix";
 import { FishbowlActiveTracker } from "../../../../backend";
-import { useTimer } from "../../hooks/useTimer";
+import { useTimer } from "../../utils/hooks/useTimer";
 import { useAdvanceWord } from "../../stateFunctions/advanceWord";
 import { useFishbowlSelector } from "../../store/fishbowlRedux";
-import { selectFishbowlPlayer } from "../../store/selectors";
+import { selectFishbowlPlayer } from "../selectors/selectors";
 import { LOW_TIME_THRESHOLD } from "../timer/FishbowlTimer";
 
 export const SomeoneGotIt = ({ timer }: { timer: FishbowlActiveTracker }) => {
