+++ packages/games/src/frontend/fishbowl/playerComponents/timer/FishbowlTimer.module.scss
